+++ mod-source-record-manager-server/src/main/java/org/folio/verticle/consumers/util/SpecificEventHandler.java
@@ -1,0 +1,13 @@
+package org.folio.verticle.consumers.util;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import org.folio.DataImportEventPayload;
+import org.folio.services.journal.JournalRecordMapperException;
+import org.folio.services.journal.JournalService;
+
+public interface SpecificEventHandler {
+
+  void handle(JournalService journalService, DataImportEventPayload eventPayload, String tenantId)
+    throws JournalRecordMapperException, JsonProcessingException;
+
+}